bug fixed

+++ Code/Utils.cpp
@@ -13,7 +13,7 @@ map<string , int> weekDays{{"Monday", 1}, {"Tuesday", 2}, {"Wednesday", 3}, {"Th
 
 bool sorterHorarioSlot(Slot* s1,Slot* s2){
     if(weekDays[s1->getDiaSemana()] != weekDays[s2->getDiaSemana()]) return (weekDays[s1->getDiaSemana()] < weekDays[s2->getDiaSemana()]);
-    if(s1->getHorarioInicio() != s2->getHorarioInicio()) return (stoi(s1->getHorarioInicio()) < stoi(s2->getHorarioInicio()));
+    if(s1->getHorarioInicio() != s2->getHorarioInicio()) return (stof(s1->getHorarioInicio()) < stof(s2->getHorarioInicio()));
     return false;
 }
